TYP: move imports inside TYPE_CHECKING
