+++ pandas/core/computation/eval.py
@@ -4,6 +4,7 @@ Top level ``eval`` module.
 from __future__ import annotations
 
 import tokenize
+from typing import TYPE_CHECKING
 import warnings
 
 from pandas._libs.lib import no_default
@@ -15,11 +16,13 @@ from pandas.core.computation.expr import (
     PARSERS,
     Expr,
 )
-from pandas.core.computation.ops import BinOp
 from pandas.core.computation.parsing import tokenize_string
 from pandas.core.computation.scope import ensure_scope
 
 from pandas.io.formats.printing import pprint_thing
+
+if TYPE_CHECKING:
+    from pandas.core.computation.ops import BinOp
 
 
 def _check_engine(engine: str | None) -> str:
