+++ pandas/core/groupby/categorical.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import numpy as np
 
 from pandas.core.algorithms import unique1d
@@ -8,7 +10,9 @@ from pandas.core.arrays.categorical import (
     CategoricalDtype,
     recode_for_categories,
 )
-from pandas.core.indexes.api import CategoricalIndex
+
+if TYPE_CHECKING:
+    from pandas.core.indexes.api import CategoricalIndex
 
 
 def recode_for_groupby(
