+++ pandas/core/indexes/extension.py
@@ -4,6 +4,7 @@ Shared methods for Index subclasses backed by ExtensionArray.
 from __future__ import annotations
 
 from typing import (
+    TYPE_CHECKING,
     Callable,
     TypeVar,
 )
@@ -21,9 +22,11 @@ from pandas.util._decorators import (
 
 from pandas.core.dtypes.generic import ABCDataFrame
 
-from pandas.core.arrays import IntervalArray
-from pandas.core.arrays._mixins import NDArrayBackedExtensionArray
 from pandas.core.indexes.base import Index
+
+if TYPE_CHECKING:
+    from pandas.core.arrays import IntervalArray
+    from pandas.core.arrays._mixins import NDArrayBackedExtensionArray
 
 _T = TypeVar("_T", bound="NDArrayBackedExtensionIndex")
 _ExtensionIndexT = TypeVar("_ExtensionIndexT", bound="ExtensionIndex")
