+++ pandas/core/interchange/dataframe.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 from collections import abc
 from typing import TYPE_CHECKING
 
-import pandas as pd
 from pandas.core.interchange.column import PandasColumn
 from pandas.core.interchange.dataframe_protocol import DataFrame as DataFrameXchg
 
 if TYPE_CHECKING:
+    import pandas as pd
     from pandas import Index
