+++ pandas/io/formats/string.py
@@ -4,12 +4,17 @@ Module for formatting output data in console (to string).
 from __future__ import annotations
 
 from shutil import get_terminal_size
-from typing import Iterable
+from typing import (
+    TYPE_CHECKING,
+    Iterable,
+)
 
 import numpy as np
 
-from pandas.io.formats.format import DataFrameFormatter
 from pandas.io.formats.printing import pprint_thing
+
+if TYPE_CHECKING:
+    from pandas.io.formats.format import DataFrameFormatter
 
 
 class StringFormatter:
