+++ pandas/io/formats/xml.py
@@ -5,7 +5,10 @@ from __future__ import annotations
 
 import codecs
 import io
-from typing import Any
+from typing import (
+    TYPE_CHECKING,
+    Any,
+)
 
 from pandas._typing import (
     CompressionOptions,
@@ -20,7 +23,6 @@ from pandas.util._decorators import doc
 from pandas.core.dtypes.common import is_list_like
 from pandas.core.dtypes.missing import isna
 
-from pandas.core.frame import DataFrame
 from pandas.core.shared_docs import _shared_docs
 
 from pandas.io.common import get_handle
@@ -28,6 +30,9 @@ from pandas.io.xml import (
     get_data_from_filepath,
     preprocess_data,
 )
+
+if TYPE_CHECKING:
+    from pandas import DataFrame
 
 
 @doc(
