+++ pandas/io/html.py
@@ -10,6 +10,7 @@ from collections import abc
 import numbers
 import re
 from typing import (
+    TYPE_CHECKING,
     Iterable,
     Pattern,
     Sequence,
@@ -30,7 +31,6 @@ from pandas.util._decorators import deprecate_nonkeyword_arguments
 from pandas.core.dtypes.common import is_list_like
 
 from pandas.core.construction import create_series_with_explicit_dtype
-from pandas.core.frame import DataFrame
 
 from pandas.io.common import (
     file_exists,
@@ -42,6 +42,9 @@ from pandas.io.common import (
 )
 from pandas.io.formats.printing import pprint_thing
 from pandas.io.parsers import TextParser
+
+if TYPE_CHECKING:
+    from pandas import DataFrame
 
 _IMPORTS = False
 _HAS_BS4 = False
