+++ pandas/io/parsers/arrow_parser_wrapper.py
@@ -1,13 +1,16 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from pandas._typing import ReadBuffer
 from pandas.compat._optional import import_optional_dependency
 
 from pandas.core.dtypes.inference import is_integer
 
-from pandas.core.frame import DataFrame
-
 from pandas.io.parsers.base_parser import ParserBase
+
+if TYPE_CHECKING:
+    from pandas import DataFrame
 
 
 class ArrowParserWrapper(ParserBase):
