+++ pandas/io/parsers/python_parser.py
@@ -10,6 +10,7 @@ import re
 import sys
 from typing import (
     IO,
+    TYPE_CHECKING,
     DefaultDict,
     Hashable,
     Iterator,
@@ -38,15 +39,16 @@ from pandas.util._exceptions import find_stack_level
 from pandas.core.dtypes.common import is_integer
 from pandas.core.dtypes.inference import is_dict_like
 
-from pandas import (
-    Index,
-    MultiIndex,
-)
-
 from pandas.io.parsers.base_parser import (
     ParserBase,
     parser_defaults,
 )
+
+if TYPE_CHECKING:
+    from pandas import (
+        Index,
+        MultiIndex,
+    )
 
 # BOM character (byte order mark)
 # This exists at the beginning of a file to indicate endianness
