+++ pandas/io/spss.py
@@ -1,15 +1,19 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Sequence
+from typing import (
+    TYPE_CHECKING,
+    Sequence,
+)
 
 from pandas.compat._optional import import_optional_dependency
 
 from pandas.core.dtypes.inference import is_list_like
 
-from pandas.core.api import DataFrame
-
 from pandas.io.common import stringify_path
+
+if TYPE_CHECKING:
+    from pandas import DataFrame
 
 
 def read_spss(
